requirements.txt
dj-database-url==0.5.0
Django==2.0.5
django-friendship==1.8.0
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
whitenoise==4.1.2
